+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.4.0] - 2018-04-13
+### Added
+- Improved error messages.
+- Plugin tested with Moodle 2.9 and 3.1.
+
 ## [0.3.0] - 2018-04-13
 ### Added
 - Initial working release.
@@ -9,4 +14,3 @@ All notable changes to this project will be documented in this file.
 ## [0.1.0] - 2018-04-12
 ### Added
 - Initial release.
-- Plugin tested with Moodle 2.9 and 3.1.